cloud_functions/es_indexer_fn /requirements.txt
google-cloud-bigquery>=2.0.0
google-cloud-secret-manager>=2.0.0
elasticsearch>=7.17.0,<8.0.0